+++ sdk/communication/azure-communication-common/src/main/java/com/azure/android/communication/common/CommunicationIdentifier.java
@@ -8,4 +8,9 @@ package com.azure.android.communication.common;
  * Common Communication Identifier for Communication Services
  */
 public abstract class CommunicationIdentifier {
+    /**
+     * Get full id of the identifier
+     * @return full id of the identifier
+     */
+    public abstract String getId();
 }