+++ sdk/communication/azure-communication-common/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## 1.0.0-beta.4 (Unreleased)
 ### Breaking Changes
-    Renamed `PhoneNumber` to `PhoneNumberIdentifier`
+   Renamed `PhoneNumber` to `PhoneNumberIdentifier`
     Renamed `CommunicationUser` to `CommunicationUserIdentifier `
     Renamed `CallingApplication` to `CallingApplicationIdentifier`
     Renamed `CommunicationUserCredential` to `CommunicationTokenCredential`
     Modified `CommunicationTokenCredential` constructor signatures to take new argument type 'CommunicationTokenRefreshOptions`
+
+### Added
+- Added `MicrosoftTeamsUserIdentifier`
 
 ## 1.0.0-beta.3 (Unreleased)